rename error-config.ts & error-config.spec.ts
